refa: remove wiki model
